requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
enum34==1.1.6
eventlet==0.22.1
Flask==0.12.2
functools32==3.2.3.post2
greenlet==0.4.13
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonpickle==0.9.6
jsonschema==2.6.0
MarkupSafe==1.0
numpy==1.14.2
python-engineio==2.0.4
python-socketio==1.9.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
